Address project catalog review feedback

+++ services/internal/project-catalog/internal/repository/postgres/project/sql/branch_rules__list.sql
@@ -7,4 +7,4 @@ WHERE project_id = @project_id
   AND (@repository_id::uuid IS NULL OR repository_id = @repository_id)
   AND (cardinality(@statuses::text[]) = 0 OR status = ANY(@statuses::text[]))
 ORDER BY pattern, id
-LIMIT @limit OFFSET @offset;
+LIMIT @limit::integer OFFSET @offset::integer;
